Add option to keep original file after adding language code
This commit expands on the work done in
14c2f1bc24efeb5fe5a67d6e11a4c4ec4185f2e3 by adding an option to
configure whether the original subtitle file is removed after the
language code is appended to the filename.

Default this to true to maintain backwards compatibility.

This option is useful when the subtitles operated on by SRT2UTF-8 are
part of torrents still being seeded. By setting the new option to false,
we don't tamper with the original subtitle file and we ensure that the
torrent checksum does not change.